requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]
python-dotenv
pypdf
google-generativeai



#https://aistudio.google.com/api-keys